TASEP-Smarticles/requirements.txt
gymnasium==0.29.1
pygame==2.5.2
pyqt6==6.6.0
scienceplots==2.1.1
tabulate==0.8.10
torchrl==0.2.1
torchvision==0.16.0
tqdm==4.66.1
